% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AppData.R
\name{AppData}
\alias{AppData}
\title{Get App data}
\description{
Get App data

Get App data
}
\details{
R6 class that allows to get the information needed for an User session of the app.
}
\section{Super classes}{
\code{\link[reportesAPI:DBManager]{reportesAPI::DBManager}} -> \code{\link[reportesAPI:User]{reportesAPI::User}} -> \code{\link[reportesAPI:Organisation]{reportesAPI::Organisation}} -> \code{\link[reportesAPI:Group]{reportesAPI::Group}} -> \code{\link[reportesAPI:Task]{reportesAPI::Task}} -> \code{AppData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AppData-new}{\code{AppData$new()}}
\item \href{#method-AppData-clone}{\code{AppData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_execute_statement"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_execute_statement'><code>reportesAPI::DBManager$db_execute_statement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_get_query"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_get_query'><code>reportesAPI::DBManager$db_get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_make_query"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_make_query'><code>reportesAPI::DBManager$db_make_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="User" data-id="user_add"><a href='../../reportesAPI/html/User.html#method-User-user_add'><code>reportesAPI::User$user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="User" data-id="user_delete"><a href='../../reportesAPI/html/User.html#method-User-user_delete'><code>reportesAPI::User$user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="User" data-id="user_edit_names"><a href='../../reportesAPI/html/User.html#method-User-user_edit_names'><code>reportesAPI::User$user_edit_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_edit"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_edit'><code>reportesAPI::Organisation$org_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_finalize"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_finalize'><code>reportesAPI::Organisation$org_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_initialize"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_initialize'><code>reportesAPI::Organisation$org_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_add"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_add'><code>reportesAPI::Organisation$org_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_delete"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_delete'><code>reportesAPI::Organisation$org_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_edit"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_edit'><code>reportesAPI::Organisation$org_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_edit"><a href='../../reportesAPI/html/Group.html#method-Group-group_edit'><code>reportesAPI::Group$group_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_finalize"><a href='../../reportesAPI/html/Group.html#method-Group-group_finalize'><code>reportesAPI::Group$group_finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_initialize"><a href='../../reportesAPI/html/Group.html#method-Group-group_initialize'><code>reportesAPI::Group$group_initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_select"><a href='../../reportesAPI/html/Group.html#method-Group-group_select'><code>reportesAPI::Group$group_select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_unit_add"><a href='../../reportesAPI/html/Group.html#method-Group-group_unit_add'><code>reportesAPI::Group$group_unit_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_unit_delete"><a href='../../reportesAPI/html/Group.html#method-Group-group_unit_delete'><code>reportesAPI::Group$group_unit_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_unit_edit"><a href='../../reportesAPI/html/Group.html#method-Group-group_unit_edit'><code>reportesAPI::Group$group_unit_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_user_add"><a href='../../reportesAPI/html/Group.html#method-Group-group_user_add'><code>reportesAPI::Group$group_user_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_user_delete"><a href='../../reportesAPI/html/Group.html#method-Group-group_user_delete'><code>reportesAPI::Group$group_user_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Group" data-id="group_user_edit"><a href='../../reportesAPI/html/Group.html#method-Group-group_user_edit'><code>reportesAPI::Group$group_user_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="progress_add"><a href='../../reportesAPI/html/Task.html#method-Task-progress_add'><code>reportesAPI::Task$progress_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="task_add"><a href='../../reportesAPI/html/Task.html#method-Task-task_add'><code>reportesAPI::Task$task_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="task_delete"><a href='../../reportesAPI/html/Task.html#method-Task-task_delete'><code>reportesAPI::Task$task_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="task_edit_metadata"><a href='../../reportesAPI/html/Task.html#method-Task-task_edit_metadata'><code>reportesAPI::Task$task_edit_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="task_get_history"><a href='../../reportesAPI/html/Task.html#method-Task-task_get_history'><code>reportesAPI::Task$task_get_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Task" data-id="task_report_progress"><a href='../../reportesAPI/html/Task.html#method-Task-task_report_progress'><code>reportesAPI::Task$task_report_progress()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppData-new"></a>}}
\if{latex}{\out{\hypertarget{method-AppData-new}{}}}
\subsection{Method \code{new()}}{
Start a session of app data based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppData$new(email = Sys.getenv("REPORTES_EMAIL"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AppData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AppData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AppData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
