% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Group.R
\name{Group}
\alias{Group}
\title{Get Group data}
\description{
Get Group data

Get Group data
}
\details{
R6 class that allows to get the Group information.
}
\section{Super classes}{
\code{\link[reportesAPI:DBManager]{reportesAPI::DBManager}} -> \code{\link[reportesAPI:User]{reportesAPI::User}} -> \code{\link[reportesAPI:Organisation]{reportesAPI::Organisation}} -> \code{Group}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groups}}{List containing the group affiliations of the User}

\item{\code{group_users}}{List containing the user list of the group The info is shown following the User's group role.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Group-new}{\code{Group$new()}}
\item \href{#method-Group-group_initialize}{\code{Group$group_initialize()}}
\item \href{#method-Group-group_edit}{\code{Group$group_edit()}}
\item \href{#method-Group-group_user_add}{\code{Group$group_user_add()}}
\item \href{#method-Group-group_user_delete}{\code{Group$group_user_delete()}}
\item \href{#method-Group-group_user_edit}{\code{Group$group_user_edit()}}
\item \href{#method-Group-group_finalize}{\code{Group$group_finalize()}}
\item \href{#method-Group-clone}{\code{Group$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_execute_statement"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_execute_statement'><code>reportesAPI::DBManager$db_execute_statement()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_get_query"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_get_query'><code>reportesAPI::DBManager$db_get_query()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="db_make_query"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-db_make_query'><code>reportesAPI::DBManager$db_make_query()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="DBManager" data-id="get_timestamp"><a href='../../reportesAPI/html/DBManager.html#method-DBManager-get_timestamp'><code>reportesAPI::DBManager$get_timestamp()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="User" data-id="user_edit_names"><a href='../../reportesAPI/html/User.html#method-User-user_edit_names'><code>reportesAPI::User$user_edit_names()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_edit"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_edit'><code>reportesAPI::Organisation$org_edit()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_finalize"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_finalize'><code>reportesAPI::Organisation$org_finalize()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_initialize"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_initialize'><code>reportesAPI::Organisation$org_initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_add"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_add'><code>reportesAPI::Organisation$org_user_add()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_delete"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_delete'><code>reportesAPI::Organisation$org_user_delete()</code></a></li>
<li><span class="pkg-link" data-pkg="reportesAPI" data-topic="Organisation" data-id="org_user_edit"><a href='../../reportesAPI/html/Organisation.html#method-Organisation-org_user_edit'><code>reportesAPI::Organisation$org_user_edit()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-new"></a>}}
\if{latex}{\out{\hypertarget{method-Group-new}{}}}
\subsection{Method \code{new()}}{
Start a Group based on an user email
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$new(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{The email the user started the session with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_initialize}{}}}
\subsection{Method \code{group_initialize()}}{
Initialize a group for a new user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_initialize(org_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_edit}{}}}
\subsection{Method \code{group_edit()}}{
Edit group metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_edit(org_id, group_id, group_title, group_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{group_title}}{The new title of the group}

\item{\code{group_description}}{The new description of the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_add"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_add}{}}}
\subsection{Method \code{group_user_add()}}{
Add an user to a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_add(org_id, group_id, user_id, group_role = "user")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{group_role}}{The role for the user in the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_delete"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_delete}{}}}
\subsection{Method \code{group_user_delete()}}{
Delete an user from a group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_delete(org_id, group_id, user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_user_edit"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_user_edit}{}}}
\subsection{Method \code{group_user_edit()}}{
Edit the role of a user inside a group and related information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_user_edit(org_id, group_id, user_id, group_role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_id}}{The id of the group on which the statement will be executed}

\item{\code{user_id}}{The id of the user on which the statement will be executed}

\item{\code{group_role}}{The role for the user in the group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-group_finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Group-group_finalize}{}}}
\subsection{Method \code{group_finalize()}}{
Remove the existence of an organisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$group_finalize(org_id, group_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_id}}{The id of the organisation on which the statement will be executed}

\item{\code{group_id}}{The id of the group on which the statement will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Group-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
